add role

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth as FacadesAuth;
 
 class CategoryController extends Controller
 {
@@ -14,7 +15,11 @@ class CategoryController extends Controller
     }
 
     public function CategoryCreate(){
-         $categories = Category::all();
+        // if(!FacadesAuth::guard('panel')->user()->can('cacreateCategory', Category::class)){
+        //     abort(403);
+        // }
+        //  $categories = Category::all();
+        $categories = Category::latest()->get();
         return view('category.category_create', compact('categories'));
     }
 
@@ -32,17 +37,19 @@ class CategoryController extends Controller
     }
 
     public function CategoryDelete($id){
+        if(!FacadesAuth::guard('panel')->user()->can('delete', Category::class)){
+            abort(403);
+        }
         $categories = Category::findOrFail($id);
         
         $categories->delete();
 
-        return redirect()->back()->with("delete", "Category Added!");
+        return redirect()->back()->with("delete", "Category Delete!");
 
     }
 
     public function categoryPageForUser(){
         $categories = Category::all();
-
         return view('category.category-wise-product', compact('categories'));
         
     }
